# packages/regpoly-web/tests/fixtures/openapi.v2.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/generators/compare": {
      "get": {
        "operationId": "v2_compare_generators_api_v2_generators_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response V2 Compare Generators Api V2 Generators Compare Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Compare Generators",
        "tags": [
          "v2"
        ]
      }
    },
    "/api/v2/generators/export": {
      "get": {
        "operationId": "v2_export_generators_api_v2_generators_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ids",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fmt",
            "required": false,
            "schema": {
              "default": "csv",
              "title": "Fmt",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Export Generators",
        "tags": [
          "v2"
        ]
      }
    },
...
    "/api/v2/generators/{gen_id}/related": {
      "get": {
        "operationId": "v2_generator_related_api_v2_generators__gen_id__related_get",
        "parameters": [
          {
            "in": "path",
            "name": "gen_id",
            "required": true,
            "schema": {
              "title": "Gen Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_RelatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Generator Related",
        "tags": [
          "v2"
        ]
      }
    },
...
    "/api/v2/library/compare": {
      "get": {
        "description": "Per-id compare payload for library entries. Library ids are\nstring-typed (not integer), and we look them up against the\nin-memory catalog. Missing ids appear as null.",
        "operationId": "v2_compare_library_entries_api_v2_library_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response V2 Compare Library Entries Api V2 Library Compare Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Compare Library Entries",
        "tags": [
          "v2"
        ]
      }
    },
...
    "/api/v2/tested-generators/compare": {
      "get": {
        "description": "Per-id compare payload for tested generators (parity with\n/api/v2/generators/compare). Missing ids appear as null.",
        "operationId": "v2_compare_tested_generators_api_v2_tested_generators_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response V2 Compare Tested Generators Api V2 Tested Generators Compare Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Compare Tested Generators",
        "tags": [
          "v2"
        ]
      }
    },
    "/api/v2/tested-generators/export": {
      "get": {
        "operationId": "v2_export_tested_generators_api_v2_tested_generators_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ids",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fmt",
            "required": false,
            "schema": {
              "default": "csv",
              "title": "Fmt",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Export Tested Generators",
        "tags": [
          "v2"
        ]
      }
    },
...
  }
}